requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
absl-py==1.4.0
addict==2.4.0
ansi2html==1.8.0
appdirs==1.4.4
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.7.0
brotlipy==0.7.0
cachetools==5.3.1
click==8.1.3
comm==0.2.0
ConfigArgParse==1.7
contourpy==1.1.0
cycler==0.11.0
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
docker-pycreds==0.4.0
executing==1.2.0
fastjsonschema==2.19.0
fire==0.5.0
Flask==3.0.0
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.20.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.14.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.4
llvmlite==0.38.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.6
mkl-service==2.4.0
nbformat==5.7.0
nest-asyncio==1.5.8
numpy==1.20.3
oauthlib==3.2.2
open3d==0.17.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==4.0.0
plotly==5.18.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
pyquaternion==0.9.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
referencing==0.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.26.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
wcwidth==0.2.6
Werkzeug==3.0.1
widgetsnbextension==4.0.9
zipp==3.15.0
 
# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
# ipython  # interactive notebook
# psutil  # system utilization
# thop  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow

